First commit
GenerateCode folder contains the EnumGenerator class. And the
EnumGeneratorWindow for Unity. Copy this to an Editor folder in your
Unity project. Now you can go to Window -> Generate Code -> Enum
Generator. You can generate enums for the tags and/or layers of your
Unity project.
The GenerateCodeExample folder contains an example console application
that uses the EnumGenerator class.